Refactor ChannelMetadata.
